PositionComponent moved to "positioning" folder
